- Removed build directory. - Removed Datasets subdirectories meant for mallet. - NGGHandler.java : Fixed casting problem in learning factor. - RunHandler.java : Added instance storage as ARFF files. - RunHandler.java : Fixed test instance list generation per fold. - TODO: Should remove mallet dependencies.

+++ src/representation/NGGHandler.java
@@ -33,7 +33,7 @@ public class NGGHandler implements GenomicSequenceRepresentationHandler<List<Doc
         
         for(int i = 0; i < representation.size()*0.9; i++) {
             for(int j = 0; j < representation.get(i).size(); j++) {
-                tempGraph.mergeGraph(representation.get(i).get(j), 1/(iCount + 1));
+                tempGraph.mergeGraph(representation.get(i).get(j), 1.0/((double)iCount + 1.0));
                 iCount++;
             }
         }
